requirements.txt
django
djangorestframework
environs
ruff
num2words
python-decouple